+++ src/commands/update-stacks.ts
@@ -65,10 +65,13 @@ export class UpdateStacksCommand extends BaseCliCommand<IUpdateStacksCommandArgs
         if (!command.stackName) {
             throw new OrgFormationError('argument --stack-name is missing');
         }
+
         Validator.validatePositiveInteger(command.maxConcurrentStacks, 'maxConcurrentStacks');
         Validator.validatePositiveInteger(command.failedStacksTolerance, 'failedStacksTolerance');
         Validator.validateBoolean(command.terminationProtection, 'terminationProtection');
         Validator.validateBoolean(command.updateProtection, 'updateProtection');
+
+        this.storeCommand(command);
 
         const terminationProtection = command.terminationProtection === true;
         const updateProtection = command.updateProtection;